requirements.txt
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.3.0
cffi==2.0.0
cryptography==45.0.7
ecdsa==0.19.1
fastapi==0.116.1
greenlet==3.2.4
idna==3.10
passlib==1.7.4
psycopg2==2.9.10
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
websockets==15.0.1
